Added scripts for smooth setup

+++ actors/broker/src/__init__.py
@@ -1,6 +1,7 @@
 # Declare Flask app for brokers
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
+import os
 
 # Declare Flask app for brokers
 app = Flask(__name__)
@@ -12,8 +13,11 @@ db = SQLAlchemy(app)
 from db_models import *
 
 with app.app_context():
-    db.drop_all()
-    db.create_all()
+    if os.getenv("DEBUG") == "true":
+        db.drop_all()
+        db.create_all()
+    else:
+        db.create_all()
 
 
 # This needs to imported, otherwise the api endpoints from views aren't integrated.
